backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
PyPDF2==3.0.1
groq==0.4.1
openai==1.3.0
anthropic==0.7.8
werkzeug==2.3.7
python-docx==0.8.11
